requirements.txt
crewai==0.19.0
langchain-community==0.0.25
pydantic==2.6.3
load-dotenv==0.1.0
crewai-tools==0.0.15